ECOCKPIT-45 imx8: cpu: force boot device instead of asking SCUFW
In ecockpit, SCU boots from SD card, but eMMC is the boot device
for Android on A72, so we hardcode the boot device 'eMMC' for A72,
so that Android bootargs are correct, and we hardcode the boot device
to 'SD2' for A53.

Ideally, boot device should be a device-tree parameter.

Signed-off-by: Fabrice Goucem <fabrice.goucem@nxp.com>