ECOCKPIT-153 leave the UART2 powered on for A72 at exit
At exit, uart2 needs to be left powered on for A72, just like uart0
is for A53.
This is required in beta3, as now all other resources are powered off
by u-boot at exit.
This hack means that device-tree flags 'debug_console' and 'early_power_on'
are not working in kernel.

Signed-off-by: Fabrice Goucem <fabrice.goucem@nxp.com>